third_party/WebKit/LayoutTests/bluetooth/characteristic/notifications/start-succeeds.html

<!DOCTYPE html>
<script src="../../../resources/testharness.js"></script>
<script src="../../../resources/testharnessreport.js"></script>
<script src="../../../resources/bluetooth/bluetooth-helpers.js"></script>
<script src="../../../resources/bluetooth/web-bluetooth-test.js"></script>
<script src="../../../resources/mojo-helpers.js"></script>
<script>
'use strict';
promise_test(() => {
  return getMeasurementIntervalCharacteristic()
    .then(({characteristic, fake_characteristic}) => {
      return fake_characteristic.setNextSubscribeToNotificationsResponse({
          gatt_code: GATT_SUCCESS})
        .then(() => characteristic.startNotifications())
        .then(start_characteristic => {
          assert_equals(start_characteristic, characteristic,
                  'Start characteristic should be the same as characteristic.');
        })
        .then(() => fake_characteristic.isNotifying())
        .then(is_notifying => assert_true(is_notifying))
        .then(() => fake_characteristic.setNextUnsubscribeFromNotificationsResponse({
          gatt_code: GATT_SUCCESS}))
    });
}, 'Single start notifications succeeds.');
</script>
